Extract logic code to seperate file, to prepare for tests
